requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
idna==3.7
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.9.1
numpy==2.0.0
packaging==24.1
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
requests==2.32.3
six==1.16.0
soupsieve==2.5
typing_extensions==4.12.2
urllib3==2.2.2
wordcloud==1.9.3